backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
alembic==1.13.1
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
bcrypt==4.1.3
pydantic-settings==2.2.1
pydantic[email]==2.7.1
redis==5.0.4
yfinance==0.2.40
pandas==2.2.2
httpx==0.27.0
python-multipart==0.0.9
pytest==8.2.0
